requirements.txt
africastalking==1.2.3
certifi==2020.12.5
chardet==4.0.0
contextlib2==0.6.0.post1
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
openpyxl==3.0.6
python-dotenv==0.15.0
requests==2.25.1
schema==0.7.4
urllib3==1.26.3